I added a cmd menu, made 2 method of use
